数据表相关开发，修改sm_field表名

+++ system-module/src/main/java/org/module/service/system/impl/db/SmFiledServiceImpl.java
@@ -2,13 +2,14 @@ package org.module.service.system.impl.db;
 
 import java.util.List;
 
-import org.module.dto.system.db.SmFiledDto;
-import org.module.mapper.system.db.SmFiledMapper;
-import org.module.model.system.db.SmFiled;
+import org.module.dto.system.db.SmFieldDto;
+import org.module.mapper.system.db.SmFieldMapper;
+import org.module.model.system.db.SmField;
 import org.module.result.EntityResult;
 import org.module.result.RootResult;
 import org.module.service.impl.BaseServiceImpl;
 import org.module.service.system.db.ISmFiledService;
+import org.module.util.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,13 +21,13 @@ import org.springframework.stereotype.Service;
  * 时间: 2017年7月12日 上午11:20:16
  */
 @Service
-public class SmFiledServiceImpl extends BaseServiceImpl<SmFiled, SmFiledMapper, SmFiledDto> implements ISmFiledService {
+public class SmFiledServiceImpl extends BaseServiceImpl<SmField, SmFieldMapper, SmFieldDto> implements ISmFiledService {
 
 	@Autowired
-	private SmFiledMapper mapper;
+	private SmFieldMapper mapper;
 
 	@Override
-	public EntityResult insertSelective(SmFiled entity) {
+	public EntityResult insertSelective(SmField entity) {
 		EntityResult result = super.insertSelective(entity);
 		if (result.getCode() == 0) {
 			mapper.insertTableColumn(entity);
@@ -35,7 +36,7 @@ public class SmFiledServiceImpl extends BaseServiceImpl<SmFiled, SmFiledMapper, 
 	}
 
 	@Override
-	public EntityResult updateByCode(SmFiled entity) {
+	public EntityResult updateByCode(SmField entity) {
 		EntityResult result = super.updateByCode(entity);
 		if (result.getCode() == 0) {
 			mapper.editTableColumn(entity);
@@ -45,11 +46,17 @@ public class SmFiledServiceImpl extends BaseServiceImpl<SmFiled, SmFiledMapper, 
 
 	@Override
 	public RootResult deleteByCode(String code) {
-		EntityResult result = super.selectByCode(code);
-		if (result.getCode() == 0) {
-			SmFiled entity = (SmFiled) result.getEntity();
-			mapper.dropTableColumn(entity);
-			super.deleteByCode(code);
+		RootResult result = new RootResult();
+		EntityResult filedResult = super.selectByCode(code);
+		if (filedResult.getCode() == 0) {
+			SmField entity = (SmField) filedResult.getEntity();
+			if (StringUtils.isNotBlank(entity.getFieldKey())) {
+				result.setCode(-1);
+				result.setMessage("字段包含索引信息，请先删除字段索引信息");
+			} else {
+				mapper.dropTableColumn(entity);
+				result = super.deleteByCode(code);
+			}
 		}
 		return result;
 	}
@@ -61,9 +68,15 @@ public class SmFiledServiceImpl extends BaseServiceImpl<SmFiled, SmFiledMapper, 
 			for (String code : codes) {
 				EntityResult entityResult = super.selectByCode(code);
 				if (entityResult.getCode() == 0) {
-					SmFiled entity = (SmFiled) entityResult.getEntity();
-					mapper.dropTableColumn(entity);
-					super.deleteByCode(code);
+					SmField entity = (SmField) entityResult.getEntity();
+					if (StringUtils.isNotBlank(entity.getFieldKey())) {
+						result.setCode(-1);
+						result.setMessage("字段包含索引信息，请先删除字段索引信息");
+						break;
+					} else {
+						mapper.dropTableColumn(entity);
+						super.deleteByCode(code);
+					}
 				}
 			}
 		} catch (Exception e) {
